Encapsulate logic
